# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/api/executions": {
      "post": {
        "tags": [
          "execution-controller"
        ],
        "summary": "Create execution",
        "operationId": "postApiExecutions",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExecutionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/executions/{id}": {
      "get": {
        "tags": [
          "execution-controller"
        ],
        "summary": "Get execution",
        "operationId": "getApiExecutionsById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/executions/{id}/events": {
      "get": {
        "tags": [
          "execution-controller"
        ],
        "summary": "Stream execution events",
        "operationId": "getApiExecutionsByIdEvents",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Last-Event-ID",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server-sent events stream",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/developers/{id}/revenue": {
      "get": {
        "tags": [
          "revenue-controller"
        ],
        "summary": "Get developer revenue",
        "operationId": "getApiDevelopersByIdRevenue",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20,
              "maximum": 100,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeveloperRevenueResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreateExecutionRequest": {
        "type": "object",
        "properties": {
          "quoteId": {
            "type": "string",
            "format": "uuid"
          },
          "maxBudgetAtomic": {
            "type": "string",
            "minLength": 1,
            "pattern": "^[0-9]+$"
          },
          "question": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 4000,
            "minLength": 1
          },
          "input": {
            "type": "null"
          }
        },
        "required": [
          "maxBudgetAtomic",
          "quoteId"
        ]
      },
      "ExecutionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "quoteId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "RUNNING",
              "COMPLETED",
              "FAILED"
            ]
          },
          "maxBudgetAtomic": {
            "type": "string",
            "pattern": "^[0-9]+$"
          },
          "reservedCostAtomic": {
            "type": "string",
            "pattern": "^[0-9]+$"
          },
          "actualCostAtomic": {
            "type": "string",
            "pattern": "^[0-9]+$"
          },
          "question": {
            "type": [
              "string",
              "null"
            ]
          },
          "input": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/JsonNode"
              },
              {
                "type": "null"
              }
            ]
          },
          "failureCode": {
            "type": [
              "string",
              "null"
            ]
          },
          "steps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExecutionStepResponse"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "actualCostAtomic",
          "createdAt",
          "id",
          "maxBudgetAtomic",
          "quoteId",
          "reservedCostAtomic",
          "status",
          "steps",
          "updatedAt"
        ]
      },
      "ExecutionStepResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "parentStepId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "agentVersionId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "CREATED",
              "PAYMENT_REQUIRED",
              "PAYMENT_SETTLED",
              "RUNNING",
              "COMPLETED",
              "FAILED"
            ]
          },
          "costAtomic": {
            "type": "string",
            "pattern": "^[0-9]+$"
          },
          "output": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/JsonNode"
              },
              {
                "type": "null"
              }
            ]
          },
          "failureCode": {
            "type": [
              "string",
              "null"
            ]
          },
          "payments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentAttemptResponse"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "agentVersionId",
          "costAtomic",
          "createdAt",
          "id",
          "payments",
          "status",
          "updatedAt"
        ]
      },
      "JsonNode": {},
      "PaymentAttemptResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "REQUIRED",
              "AUTHORIZED",
              "SETTLED",
              "FAILED",
              "RECONCILIATION_REQUIRED"
            ]
          },
          "amountAtomic": {
            "type": "string",
            "pattern": "^[0-9]+$"
          },
          "mode": {
            "type": "string",
            "enum": [
              "simulated",
              "x402"
            ]
          },
          "transactionHash": {
            "type": [
              "string",
              "null"
            ]
          },
          "paymentIdentifier": {
            "type": [
              "string",
              "null"
            ]
          },
          "failureCode": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "amountAtomic",
          "id",
          "mode",
          "status"
        ]
      },
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          },
          "traceId": {
            "type": "string"
          }
        },
        "required": [
          "error",
          "traceId"
        ]
      },
      "ErrorBody": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/JsonNode"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "DeveloperRevenueResponse": {
        "type": "object",
        "properties": {
          "developerId": {
            "type": "string",
            "format": "uuid"
          },
          "totalRevenueAtomic": {
            "type": "string",
            "pattern": "^[0-9]+$"
          },
          "directRevenueAtomic": {
            "type": "string",
            "pattern": "^[0-9]+$"
          },
          "dependencyRevenueAtomic": {
            "type": "string",
            "pattern": "^[0-9]+$"
          },
          "directCount": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "dependencyCount": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RevenueEntryResponse"
            }
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          }
        },
        "required": [
          "dependencyCount",
          "dependencyRevenueAtomic",
          "developerId",
          "directCount",
          "directRevenueAtomic",
          "entries",
          "totalRevenueAtomic"
        ]
      },
      "RevenueEntryResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "executionStepId": {
            "type": "string",
            "format": "uuid"
          },
          "paymentAttemptId": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "enum": [
              "DIRECT",
              "DEPENDENCY"
            ]
          },
          "amountAtomic": {
            "type": "string",
            "pattern": "^[0-9]+$"
          },
          "paymentMode": {
            "type": "string",
            "enum": [
              "simulated",
              "x402"
            ]
          },
          "transactionHash": {
            "type": [
              "string",
              "null"
            ]
          },
          "paymentIdentifier": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "amountAtomic",
          "createdAt",
          "executionStepId",
          "id",
          "paymentAttemptId",
          "paymentMode",
          "type"
        ]
      },
...
    }
  }
}